+++ src/FESpaces/Pullbacks.jl
@@ -53,6 +53,19 @@ function get_cell_pushforward(
   return CoVariantPiolaMap(), change, (Jt,)
 end
 
+# DoubleContraVariantPiolaMap
+
+using Gridap.ReferenceFEs: DoubleContraVariantPiolaMap
+function get_cell_pushforward(
+  ::DoubleContraVariantPiolaMap, model::DiscreteModel, cell_reffe, conformity
+)
+  cell_map = get_cell_map(get_grid(model))
+  Jt = lazy_map(Broadcasting(∇),cell_map)
+  change = lazy_map(r -> Diagonal(ones(num_dofs(r))), cell_reffe)
+  #change = get_sign_flip(model, cell_reffe)
+  return DoubleContraVariantPiolaMap(), change, (Jt,)
+end
+
 # NormalSignMap
 
 """
@@ -91,14 +104,16 @@ function evaluate!(cache,k::NormalSignMap,reffe,facet_own_dofs,cell)
 
   setsize!(dof_sign_cache, (num_dofs(reffe),))
   dof_sign = dof_sign_cache.array
-  fill!(dof_sign, one(eltype(dof_sign)))
+
+  o = one(eltype(dof_sign))
+  fill!(dof_sign, o)
 
   facets = getindex!(cell_facets_cache,cell_facets,cell)
   for (lfacet,facet) in enumerate(facets)
     owner = facet_owners[facet]
     if owner != cell
       for dof in facet_own_dofs[lfacet]
-        dof_sign[dof] = -1.0
+        dof_sign[dof] = -o
       end
     end
   end