+++ src/ReferenceFEs/NedelecRefFEs.jl
@@ -15,7 +15,7 @@ Singleton of the [`Nedelec`](@ref) reference FE name.
 """
 const nedelec = Nedelec()
 
-Pushforward(::Type{<:Nedelec}) = CoVariantPiolaMap()
+Pushforward(::Type{Nedelec}) = CoVariantPiolaMap()
 
 """
     NedelecRefFE(::Type{T}, p::Polytope, order::Integer)